Rename ratingImages to ratingIcons
