--- src/frontend/shared/ratingImages.js
@@ -1,7 +1,0 @@
-export default {
-  3: "/assets/icons/icon_rating_3.svg",
-  2: "/assets/icons/icon_rating_2.svg",
-  1: "/assets/icons/icon_rating_1.svg",
-  0: "/assets/icons/icon_rating_0.svg",
-  null: "/assets/icons/icon_rating_na.svg",
-};
